refactor: :recycle: break down the old `pipeline.ts` file to multiple modules
Extracted functions and types to separated files to improve maintainability of this experimental feature
