+++ src/helpers/Experimental/pipeline/createPipeline.ts
@@ -1,0 +1,3 @@
+import { pipeline } from './helpers/pipeline'
+
+export const createPipeline = () => pipeline()
