+++ src/helpers/Experimental/pipeline/types/Depipe.ts
@@ -1,0 +1,1 @@
+export type Depipe<T> = (this: any) => T
