+++ src/helpers/Experimental/pipeline/types/GetAsyncPipeline.ts
@@ -1,0 +1,3 @@
+import type { GetPipeline } from './GetPipeline'
+
+export type GetAsyncPipeline<T> = GetPipeline<Promise<T>>
