+++ src/helpers/Experimental/pipeline/types/GetPipeline.ts
@@ -1,0 +1,3 @@
+import type { Pipable } from './Pipable'
+
+export type GetPipeline<T> = T & Pipable<T>
