+++ src/helpers/Experimental/pipeline/types/Pipable.ts
@@ -1,0 +1,9 @@
+import type { BaseAsyncPipable } from './interfaces/BaseAsyncPipable'
+import type { BasePipable } from './interfaces/BasePipable'
+import type { Unpipable } from './Unpipable'
+
+export type Pipable<T> = T extends Unpipable
+    ? T
+    : T extends Promise<any>
+    ? BaseAsyncPipable<T>
+    : BasePipable<T>
