+++ src/helpers/Experimental/pipeline/types/Unpipable.ts
@@ -1,0 +1,1 @@
+export type Unpipable = null | void | undefined | never
