+++ src/helpers/Experimental/pipeline/types/WithDepipe.ts
@@ -1,0 +1,3 @@
+import { HasDepipe } from './interfaces/HasDepipe'
+
+export type WithDepipe<TValue> = TValue & HasDepipe<TValue>
