+++ src/helpers/Experimental/pipeline/types/WithPipe.ts
@@ -1,0 +1,3 @@
+import { HasPipe } from './interfaces/HasPipe'
+
+export type WithPipe<TValue> = TValue & HasPipe<TValue>
