+++ src/helpers/Experimental/pipeline/types/WithPipeAsync.ts
@@ -1,0 +1,3 @@
+import { HasPipeAsync } from './interfaces/HasPipeAsync'
+
+export type WithPipeAsync<TValue> = TValue & HasPipeAsync<TValue>
