+++ src/helpers/Experimental/pipeline/types/interfaces/BaseAsyncPipable.ts
@@ -1,0 +1,4 @@
+import type { BasePipable } from './BasePipable'
+import type { HasPipeAsync } from './HasPipeAsync'
+
+export interface BaseAsyncPipable<T> extends BasePipable<T>, HasPipeAsync<T> {}
