+++ src/helpers/Experimental/pipeline/types/interfaces/BasePipable.ts
@@ -1,0 +1,4 @@
+import type { HasDepipe } from './HasDepipe'
+import type { HasPipe } from './HasPipe'
+
+export interface BasePipable<T> extends HasPipe<T>, HasDepipe<T> {}
