+++ src/helpers/Experimental/pipeline/types/interfaces/HasDepipe.ts
@@ -1,0 +1,5 @@
+import { Depipe } from '../Depipe'
+
+export interface HasDepipe<T> {
+    readonly depipe: Depipe<T>
+}
