+++ src/helpers/Experimental/pipeline/types/interfaces/HasPipe.ts
@@ -1,0 +1,5 @@
+import { Pipe } from '../Pipe'
+
+export interface HasPipe<T> {
+    readonly pipe: Pipe<T>
+}
