+++ src/helpers/Experimental/pipeline/types/interfaces/HasPipeAsync.ts
@@ -1,0 +1,5 @@
+import { AsyncPipe } from '../AsyncPipe'
+
+export interface HasPipeAsync<T> {
+    readonly pipeAsync: AsyncPipe<Awaited<T>>
+}
